[Enh]: Description `undefinedValue` Property
This allows integration with e.g. the Null Pattern. For example, let's say that my domain has a name with a single-option relation to a generational (e.g. Jr, Sr). Since generationals are optional, there is NullGenerational to avoid nil checks everywhere. Before this fix, there was no way to inject that class instead of nil. Now, there is. See the test (#testUndefinedValue).