[Bug]: Bloc Form To-One Tokens Pointing to Copy
As a result of #296, memento fields are copies of the model state. Bloc tokens were not updated to this new strategy, and thus were pointing to these copies instead of the actual state. When they were clicked, an inspector would open on a temporary copy, not the object actually held by the model.