The only way to have non-identical Date on GemStone is to change it
MADateDescriptionTest validates conditions also by validating against
non-identical instance.  The problem is that Pharo and GemStone have
different approaches to Date type.

Pharo has a Date as wrapper over DateAndTime, which enables it to use
aDate copy to create new instance of a Date.

In GemStone the Date is an immediate type.  It gives SmallDate if
possible. The only way to have a non-identical instance is to change
the date.